hardware/imuService/requirements.txt
mpu6050-raspberrypi==1.2
websockets==12.0
